feat: enhance application with new components and configuration updates
- Added Header component for navigation across the application.
- Introduced Dashboard and Farms islands for displaying agricultural data and managing farms.
- Updated deno.json to include linting rules for better code quality.
- Enhanced app routes to utilize new components for improved user experience.
